+++ test/passing/gen/gen.ml
@@ -50,7 +50,7 @@ let add_test ?base_file map src_test_name =
 let register_file tests fname =
   match String.split_on_char '.' fname with
   | test_name
-    :: (("ml" | "mli" | "mlt" | "mld" | "eliom" | "eliomi") as ext)
+    :: (("ml" | "mli" | "mlt" | "mld" | "mll" | "eliom" | "eliomi") as ext)
     :: rest -> (
       let fname = input_dir ^ fname in
       let src_test_name = test_name ^ "." ^ ext in